Mods with a Version Checker `modThreadId` in the format of `1234.0` were being read as `12340` (e.g. Detailed Combat Results).

+++ lib/jre_manager/jre_entry.dart
@@ -636,7 +636,7 @@ class CustomJreDownloadState with CustomJreDownloadStateMappable {
 class CustomJreVersionCheckerFile with CustomJreVersionCheckerFileMappable {
   final String masterVersionFile;
   final String modName;
-  final int? modThreadId;
+  final double? modThreadId;
 
   @MappableField(hook: VersionHook())
   final Version modVersion;
